@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
 import '../page/export_page_pkg.dart';
 import '../tools/global.dart';
@@ -122,7 +123,10 @@ Widget backBTN(
         width: 22.0,
         height: 22.0,
       ),
-      onPressed: () => pop(context),
+      onPressed: () {
+        SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
+        pop(context);
+      },
     ),
   );
 }
