@@ -1,17 +1,17 @@
 import 'package:json_annotation/json_annotation.dart';
-part 'eye_video.g.dart';
+part 'eye_related.g.dart';
 
 @JsonSerializable()
-class EyeVideo {
-  EyeVideo(this.itemList, this.nextPageUrl);
+class EyeRelated {
+  EyeRelated(this.itemList);
 
-  final List<Item> itemList; // 视频列表
-  final String nextPageUrl; // 每页末尾的查询下一页链接
+  final List<Item> itemList; // 推荐视频列表
 
-  factory EyeVideo.fromJson(Map<String, dynamic> json) =>
-      _$EyeVideoFromJson(json);
+  factory EyeRelated.fromJson(Map<String, dynamic> json) =>
+      _$EyeRelatedFromJson(json);
 
-  Map<String, dynamic> toJson(EyeVideo instance) => _$EyeVideoToJson(instance);
+  Map<String, dynamic> toJson(EyeRelated instance) =>
+      _$EyeRelatedToJson(instance);
 }
 
 @JsonSerializable()
@@ -19,7 +19,7 @@ class Item {
   Item(this.data, this.type);
 
   final Data data; // 集合中的条目对象
-  final String type; // 组别分类(分隔线、图片、链接)
+  final String type; // 单元分类
 
   factory Item.fromJson(Map<String, dynamic> json) => _$ItemFromJson(json);
 
@@ -28,37 +28,15 @@ class Item {
 
 @JsonSerializable()
 class Data {
-  Data(this.content);
-
-  final Content content;
-
-  factory Data.fromJson(Map<String, dynamic> json) => _$DataFromJson(json);
-
-  Map<String, dynamic> toJson(Data instance) => _$DataToJson(instance);
-}
-
-@JsonSerializable()
-class Content {
-  Content(this.data);
-
-  final InnerData data;
-
-  factory Content.fromJson(Map<String, dynamic> json) =>
-      _$ContentFromJson(json);
-
-  Map<String, dynamic> toJson(Content instance) => _$ContentToJson(instance);
-}
-
-@JsonSerializable()
-class InnerData {
-  InnerData(
+  Data(
     this.id,
     this.title,
+    this.description,
     this.author,
     this.cover,
     this.playUrl,
-    this.description,
     this.duration,
+    this.date,
   );
 
   final int id;
@@ -68,34 +46,27 @@ class InnerData {
   final Cover cover;
   final String playUrl;
   final int duration;
+  final num date;
 
-  factory InnerData.fromJson(Map<String, dynamic> json) =>
-      _$InnerDataFromJson(json);
+  factory Data.fromJson(Map<String, dynamic> json) => _$DataFromJson(json);
 
-  Map<String, dynamic> toJson(InnerData instance) =>
-      _$InnerDataToJson(instance);
+  Map<String, dynamic> toJson(Data instance) => _$DataToJson(instance);
 }
 
 @JsonSerializable()
 class Author {
   Author(this.icon, this.name, this.description);
-
   final String icon;
   final String name;
   final String description;
-
   factory Author.fromJson(Map<String, dynamic> json) => _$AuthorFromJson(json);
-
   Map<String, dynamic> toJson(Author instance) => _$AuthorToJson(instance);
 }
 
 @JsonSerializable()
 class Cover {
   Cover(this.detail);
-
   final String detail;
-
   factory Cover.fromJson(Map<String, dynamic> json) => _$CoverFromJson(json);
-
   Map<String, dynamic> toJson(Cover instance) => _$CoverToJson(instance);
 }
